major bugfix: I forgot to select the criteria properly when simulating
I only came across this when noticing that median RTs between easy and hard
conditions differed only when simulating one data set at a time, but not when
using a repeated data set (to estimate the response distribution for a given
experiment with greater precision). For some reason numba didn't complain when
I indexed into the criteria array beyond the existing elements.

This bug also meant that inference with the model shouldn't have worked,
because the criteria would not fit to the simulated trials.

Now everything should work.