script testing parameter recovery with the model and EP-ABC
I tried to simulate an experiment close to how it happened in the real
experiment. In the committed example I also set parameter values so that
summary measures roughly correspond to what Dragan observed in the experiment:
high accuracy with small difference in accuracy between easy and hard trials,
but larger effect of difficulty on (median) RTs with overall RTs around 700 ms.
I don't think that my parameter setting reproduces the actual observations, but
it goes into the right direction.

Parameter recovery appears to be quite good even though I tried to infer bound,
noisestd and intstd all at once. There is a strong posterior correlation
between bound and intstd, but this does not prevent the posterior from
narrowing close to the true values. I do observe, however, a slight
overestimation of the bound with a simultaneous underestimation of noisestd and
intstd. I suspect that parameter recovery still works with all three decision
parameters varying freely, because the data mixes easy and hard conditions.