new parameter: choice bias (prior probability of clockwise rotation)
I already had implemented a prior over directions. This could in principle
also implement a bias towards clockwise or anti-clockwise rotations, but this
would not be consistent across trials, because the definition of clockwise
depends on the trial-specific criterion orientation. So I now introduced a
higher level choice bias directly manipulating the probability that a clockwise
rotation is the correct one.

I have verified that the mechanism works: bias below 0.5 increases the fraction
of anti-clockwise responses and at the same time makes anti-clockwise responses
faster and the other way around for bias above 0.5.

Depending on the value of the bound bias values should be relatively close to
0.5. E.g., for bound=0.6 bias=0.42 already has a very strong effect.