script for testing inference on single subjects
Goes through the whole data preparation and inference procedures and produces
diagnostic plots of the results.

I tested for subject 18. Inference took 3h 45min, when I included the new bias
parameter (without it inference was a bit faster, but not much). Results appear
to be fine: The resulting response distributions cover the data distributions
quite well and exhibit a similar difference in median RTs between easy and hard
trials. Non-decision times were estimated to be pretty small with most values
within 100 to 200 ms. So most variance in RTs is actually explained with the
decision mechanism (dts ranging roughly within 250 to 1150 ms).

There are two quite strong correlations in the posterior: intstd and bound are
very strongly anti-correlated and ndtmean and ndtspread are strongly
correlated. The latter is ok meaning that only small ndts can be used to
explain the data, because the median ndt should remain roughly similar, if both
ndtmean and ndtspread are increased for small values of ndtmean. The
correlation between intstd and bound is a bit more worrisome. I would usually
interpret this as a redundancy between these two parameters which is expected
from theory and previous models, but the correlation did not lead to a strong
widening of the posterior, i.e., both intstd (within 0.52 and 0.555) and bound
(within 0.53 and 0.54) were inferred relatively precisely in comparison to the
prior. In such a case I wouldn't think that the high posterior correlation is
an issue, but it could still be that EP-ABC just estimated the posteriors too
narrowly despite the redundancy between the parameters. It could also be,
however, that in this model intstd and bound are not completely redundant, when
the model tries to explain responses across easy and hard conditions.

I originally introduced the bias parameter to capture a subject's bias to
preferentially press one of the buttons. That's also what the bias does, but it
additionally shifts the response times of the preferred choice to earlier
responses. The data of subject 18, on the other hand, shows a small bias in
choices without a visible offset of response time distributions. Consequently,
even with bias parameter the model cannot really explain the small choice bias
of the subject, at least it seems so, when I assume that inference generally
works.