effect in error rate difference driven by hard trials
I checked whether the difference in error rate between slow and fast trials is
specific for either easy or hard trials. It appears to be specific for hard
trials as there is no effect at all for easy trials, but in hard trials the
effect size is approximately the same as when all trials are included.

I also excluded timed out trials in this analysis. So these could not have an
effect (and anyways didn't, because they were encoded as nan in the error
data).