making angle defintions consistent between model and data + bugfix
According to the stored data Dragan used an angle definition different from the
one I used for the model. In the model positive angles denote anti-clockwise
rotation while in Dragan's data positive angles correspond to clockwise
rotation. So, when loading the data for use with the model, I now flip the sign
of the loaded angles to make them consistent with the model.

I actually have an assert statement which is supposed to check for these
inconsistencies (by checking whether the definition of an error in the response
is consistent between model and data), but this didn't throw an exception,
because I had also flipped the definitions of the choices in my model
definition: I used -1 for clockwise and 1 for anti-clockwise response although
the meaning in the data was the opposite. So I had two errors which cancelled
each other in the model and anyway lead to correct inference.

Now everything should be consistent, but the numerical values of the angles
stored in the data txt files differ from the ones loaded for model inference.