comparing inference with fixed vs. variable intstd for subject 18
I was worried that the tight posteriors for intstd and bound despite their
strong correlation resulted only from deficiencies in EP-ABC. So I pitted two
runs of inference against each other, one repeating the previous full inference
and one in which I fixed intstd to a value (0.3) that should lead to bound
values greater than 0.6. At the same time I increased the width of the noisestd
and intstd priors a bit, because I was also worried that they either biased the
stds too much to small values and/or did not allow small enough values. Results
indicate that EP-ABC works fine with logml of the full model being larger than
the logml of the reduced model with fixed intstd (-1126 vs. -1135). However, it
turned out that with the new priors the optimal value of intstd appears to be
close to 0.3 by accident, at least that's what the posterior suggests (mean of
~0.295). So I did not end up testing whether there exist values of bound for
any setting of intstd that lead to similarly good explanations of the data,
i.e., whether intstd and bound are indeed redundant parameters in the model
with respect to the data. But that the full model wins the model comparison
even though the reduced model had almost optimal intstd already indicates that
bound and intstd (and perhaps noisestd) are not fully redundant.

Rationale for new priors
------------------------
The lognormal distribution is peculiar. Even though it has median of 1 with
mu=0, it may have a clear mode at considerably smaller values. Initially I was
worried that this introduces a strong bias towards the mode (at 0.37 for
sigma=1), but checking the quantiles of the lognormal prior I also realised
that the distribution is relatively narrow (2.5% percentile at 0.141 for
sigma=1). Thus, instead of trying to move the mode to larger values I decided
to rather just widen the prior and not care about the mode. The reasoning
behind this is that still 50% of the probability mass will be above and below 1
with mu=0. So at least EP-ABC will equally frequently encounter 'small' and
'large' values, when sampling from the prior, but I'm still not 100% sure about
the effect of the mode. I tried to get an intuition about this by looking at
the gamma distribution as a prior for a precision of a Gaussian distribution
with known mean. Then the gamma is a conjugate prior and everything is easy to
compute. The gamma can also be skewed, but it has much lighter tails than the
lognormal. So the gamma has considerably smaller difference between the
position of the mode and the median. This makes it unfortunately no good model
to get an intuition for the lognormal.

Anyway, I stuck with mu=0 and increased sigma to 1.2 so that the lognormal
covers a greater range of values.

More details about results
--------------------------
I had the two inferences run in parallel. They took 3 h 54 min and 4 hours to
compute. The reduced model was the slower one even though it had one less
parameter.

In the full model I see the strong anti-correlation between intstd and bound as
before. Just that now the bound values are around 0.7 and intstd around 0.295.
The bias has flipped from 0.499 to 0.506 and non-decision times have become a
bit slower with ndtmean=-1.4. The model now predicts slightly higher median RTs
than present in the data. It still does not completely reproduce the shape of
the response distribution although the distributions are quite similar.

The reduced model had fixed intstd=0.3. Estimated bound values were also around
0.7, but the posterior was a bit narrower than in the full model and favoured
values above 0.7 (ranges [0.7, 0.71] vs [0.688, 0.712]). The posterior for the
bias was a bit wider (ranges [0.502, 0.51] vs [0.505, 0.507]) and the
non-decision time was lower on average (ndtmean ~-1.65). The predicted median
RTs are a bit more off from the data than for the full model, but not much.
Response distributions look otherwise quite similar.

Goodness of fit measures:
   [full model] ppl-sum =  -5816.0, logml =  -1126.4
[reduced model] ppl-sum =  -5830.3, logml =  -1135.1

Conclusion
----------
I have some confidence that inference works fine with the model and that the
data allows to discriminate all parameters. As further confirmation I have the
reduced model running with intstd=0.2 now. This should lead to lower posterior
predictive likelihoods and marginal likelihood, if everything is as I now
expect.

I did not expect that my change in prior for intstd and noisestd leads to the
observed relatively large change in intstd posterior values. I should further
investigate the role of these lognormal priors.