continued test of inference for subject 18: redundancy!
My previous test was inconclusive, because with the new priors the inferred
intstd of the full model was by accident close to the value that I used in the
reduced model. The additional run with intstd=0.2 now reveals problems with the
inference and suggests that bound and intstd are indeed redundant.

I chose intstd=0.2, because that was clearly outside the 95% posterior
probability interval of the full model posterior which was roughly [0.287,
0.304]. So, if the full model posterior reflected the true posterior, or at
least the values of intstd that lead to the best fit of the model to the
responses, then setting intstd=0.2 should lead to worse posterior predictive
and marginal likelihoods. This, however, was not the case. In contrary, the
newly estimated log-likelihoods were better than those of the full model:

   [full model] ppl-sum =  -5816.0, logml =  -1126.4
[reduced model] ppl-sum =  -5782.3, logml =  -1090.3

As the strong anti-correlation of bound and intstd in the full model posterior
indicated, this result, therefore, shows that intstd and bound are indeed
redundant (at least in the range of values I tested here). Also, the result
shows that the posteriors are too narrow, because they do not capture all
values that lead to similarly good explanations of the data. This is an old
problem of EP-ABC that I thought I could get around by setting nacc=2000, as I
used here. The large number of data points here (2000) appears to lead to too
narrow posteriors despite the relatively large nacc. I will need to find a
solution for this.

Details of inference results
----------------------------
posterior probability intervals:
noisestd  [ 0.903,  0.932]
bound     [ 0.849,  0.856]
bias      [ 0.513,  0.527]
ndtmean   [-1.358, -1.314]
ndtspread [ 0.228,  0.275]
lapseprob [ 0.016,  0.022]

The most striking difference to the full model is the higher bound which
compensates for the lower intstd. The noisestd posterior is actually quite
similar to the one here, just that the one here is slightly shifted towards
smaller values (full model: [0.914, 0.947]). The bias and ndtmean are a bit
larger than in the full model.

The model response distributions do indeed look like they would fit the data
distributions a bit better than the full model. Also, the median RTs appear to
be closer to the data median RTs. Non-decision time distribution is roughly
between the very narrow one of the reduced model with intstd=0.3 and the wider
distribution of the full model.