first example of using SNL for parameter inference from subject data
This appears to work nicely. It runs roughly 40 min for subject 19. Other than
EP-ABC SNL so far uses summary statistics to infer parameter posteriors. I used
accuracy and RT quantiles for easy and hard trials as summary statistics.

The log file with the training losses of the network suggest that the network
had converged after the 7th round of learning. The resulting parameter
posterior looks plausible with slight bimodality of noisestd and intstd
posteriors (there is a side-plateau on the peaks). intstd and bound exhibit a
clear correlation, but specifically the bound still exhibits a clear preference
for values around 0.55.

Compared to previous EP-ABC results I noticed that the lapseprob is estimated
to be much higher than before with a peak around 0.1 and values reaching 0.3 in
the posterior samples. I wonder whether this is a consequence of using the
RT quantiles as summary statistics instead of inferring from the full data. My
intuition is that SNL doesn't get to know that there are regions of RT space
without any data when it is only fed the quantiles.

I might try to find a way of using single trials in SNL and wonder what a good
measure is for deciding whether one or the other approximate inference is
better.