testing variations of summary statistics
As the inference appeared to overestimate the contribution of lapses in the
data, I have explored different summary statistics.

histogram stats
---------------
I thought that the reason behind the high posterior lapse probabilities was
that the inference doesn't see RT regions withouth any data, but only the
quantile values, when using the standard quantile stats. Relatively high lapse
probabilities, therefore, do not get penalised for covering data free space. So
I imagined that I could give this information to the inference by using density
values defined by the data over fixed regions of data space as summary
statistics, i.e., by counting the data within RT bins.

However, I still wanted to make the summary statistics most informative about
the data. So I decided to implement varying bin widths with narrow bins around
the median RT and increasingly wider bins towards the edges of the data space.
Because I also didn't want to increase the dimensionality of the summary
statistics much, I decided to use only 7 bins.

This choice of bins in the end countered my original motivation of letting the
inference know of bins without data, because there were no bins without data
with this choice of bins, although there was only very little data with
correspondingly low density for the first bin from RT=0. In the end lapse
probabilities were indeed estimated lower than for the quantile stats (all
lapseprob < 0.1 with peak around 0.06). I now also think that I should be able
to get a similar effect by simply including the 0.01 quantile into the quantile
stats instead of the 0.05 quantile.

In contrast to the  lapse probability, the decision parameters (noisestd,
intstd, bound) have been estimated less precisely. For example the posterior
for noisestd reaches 0.7, the one for intstd reaches beyond 1 and posterior
bound values reach 0.9). The summary statistics values for each sample from the
posterior vary strongly and especially are not concentrated around the data
values. This can also be seen in the posterior predictive RT distributions
which are wider than the RT distributions of the data. Also, the inference has
estimated a relatively strong bias with peak around 0.43 leading to
qualitatively quite different distributions than found in the data.

Overall the histogram summary statistics did not lead to trustworthy results.
Perhaps I would have to increase the number of bins, or make equally sized bins
to make this work.

quantile stats with time outs
-----------------------------
Seeing that it was an issue to exclude time outs from the summary statistics I
have simply added the proportion of timeouts to it. This has led to more
precise estimates of lapse probabilities that are also smaller than before with
a peak around 0.07 and values reaching roughly up to 0.13. Overall the fit to
the summary statistics was pretty good, also with good correspondence between
data and posterior predictive RT distributions. I did see a pretty strong
anti-correlation between ndtmean and intstd here (-0.91), but this only led to
a widening of the ndtmean posterior towards smaller values.

quantile stats starting with 0.01 quantile
------------------------------------------
As mentioned above, I tested the quantile stats starting with the 0.01 quantile
going to the 0.95 quantile as before. I have used one more quantile per
condition (8 instead of 7) meaning that the summary statistics was 20D instead
of 18D. The computation time was a few seconds slower than before, but this was
well within the range of variability that I would expect from different runs of
the same inference (43:01 min vs. 42:45 min). The posterior predictive check
indicates a slightly worse fit than the previous run with quantile stats,
specifically for the hard trials. The lapse probability was estimated to be
slightly lower, but overall very similar to before (peak around 0.07, range up
to ~0.12). So including the small quantile had no mentionable effect on the
lapse probabilities. The strongest effect I saw was in a reduction in noisestd.
The median dropped from 0.22 to 0.16, but specifically the tail of noisestd was
cut so that it stayed under 0.4 in the posterior. Apart from that the results
were very similar to the previous run.

quantile stats in 0.01 to 0.99
------------------------------
All the same as above, just with 0.99 as last quantile instead of 0.95 (still
8 quantiles). Fit, as measured by posterior predictive RT dists, is quite
similar to before. The fit of the summary statistics suggests that the model
has difficulties in the tails of the distribution. The lower tail tends to be
overestimated (more low values than in the data) while the upper tail tends
to be underestimated (fewer high values than in the data) while the
intermediate quantiles are fit with reasonable accuracy. The lapse
probabilities are a bit higher again (peak at 0.1 range up to 0.18). This may
be the inference's attempt to get more values in the upper tail. In turn, the
estimates of the decision parameters increased a bit again, especially in the
tails.

Here's an overview of median estimates of decision parameters for the three
runs with quantile stats:

name      noisestd       intstd        bound
05-95     0.224272     0.151166     0.649528
01-95     0.157001     0.158947     0.695998
01-99     0.163679     0.169530     0.602707

and here the 0.95 quantile:

name      noisestd       intstd        bound
05-95     0.352494     0.366415     0.823569
01-95     0.318114     0.309982     0.890069
01-99     0.389277     0.323588     0.768354

It is so far unclear whether these variabilities across runs are systematic
results of the different quantiles in the summary statistics, or whether this
could also be the normal variability of the approximate inference.