re-run; model tends to explain RTs by non-decision time
I conducted a pure re-run of the last inference with 0.01-0.99 quantile
statistics. The results varied more strongly than in the previous runs with
different choice of quantiles. Specifically, the lapse probabilities were
estimated to be considerably lower (peaked close to 0 and range up to 0.06).
Additionally, the stds were slightly higher than in previous runs and the
ndtmean was lower.

name      noisestd       intstd        bound      ndtmean
5%        0.170327     0.207114     0.508489    -1.292224
50%       0.337936     0.419469     0.553729    -1.112432
95%       0.634324     0.786526     0.676429    -0.978746

These changes mean that the model explains more of the RTs using the decision
process instead of the non-decision process. Checking the dt vs ndt
distributions directly also revealed that the previous runs with ndtmean ~ -0.8
explain pretty much all of RTs with non-decision time. Now I obviously wonder
whether this is the genuine best explanation of the data, or whether the
inference is unreliable.

I can remove the model's ability to explain RTs with non-decision time by
changing the ndt-distribution to uniform instead of log-normal. That is also
well-motivated by the use of uniform distribution with the DDM.

Given the results of the last run, in which the quantiles are fitted worse than
in previous runs (tend to be underestimated), I would expect that overall the
quantile fits are worse, too, when the ndt-distribution changes to a uniform.