uniform non-decision time
I implemented the option to switch between lognormal and uniform non-decision
time distributions. To do this I renamed ndtmean to ndtloc. For both types of
distribution ndtloc will be exp-transformed before using it to determine the
location of the distribution (median for lognormal, lower bound for uniform).
This transform ensures that the location is alway positive and that the
numerical interpretation is comparable (high negative values mean small ndt).
In the uniform distribution ndtspread determines the upper bound in relation to
the given lower bound.

Note that in the plot produced by model.plot_dt_ndt_distributions the
non-decision time distribution may not look uniform even though it is, because
the bins might be too wide for a narrow ndt distribution.

Results of a run of inference are not particularly good. The RT quantiles are
consistently underestimated. More problematically, the posterior distributions
for the decision parameters have become wider and the stds appear to be
multimodal.

name      noisestd       intstd        bound       ndtloc
5%        0.128747     0.097186     0.506352    -2.276123
50%       0.669779     0.425252     0.542286    -1.758413
95%       1.487466     1.176705     0.765289    -1.277609

The good thing is that the decision model does indeed contribute more than
before, but the found solution is not convincing. Using the estimated mode of
the posterior to generate decisions almost exclusively generates choices of
one kind. This is most likely because the chosen bias is strong in relation
to the other parameters. Anyway, it is a bad sign that such a solution is
anywhere close to being the most plausible. There is of course an issue with
correlation among parameters so that the independent marginal modes are not
a good estimate of the joint mode. The correlations between bound, noisestd and
intstd are high so that it appears reasonable to actually fix one of them.