test with intstd fixed to 1
name         bound    lapseprob       ndtloc     noisestd
5%        0.501339     0.001706    -2.864768     0.094350
50%       0.504212     0.009799    -2.214703     0.536503
95%       0.511887     0.029897    -1.764947     2.464952

The training diagnostics was promising in that convergence appeared to be
faster than with free intstd and in that validation losses reached below -90
for the first time, but already the estimated log-likelihood values were lower
than before suggesting a worse fit (assuming that the likelihood function was
reasonably approximated). And the fit was pretty bad with consistent
under-estimation of RT quantiles, but even worse, a complete failure to
replicate accuracies. The problem was the extremely low bound which in
combination with a rather variable bias meant that decisions were made
in the first time step and variability of RT only resulted from non-decision
time.

Considering the previous results, it was rather stupid to choose intstd=1,
because the previous posteriors indicated that this value was very unlikely
given the data. If, however, the inference would have not worked properly,
because of the strong correlations in the posterior, the exact value of intstd
might have been irrelevant. So finding that intstd=1 is useless supports the
fidelity of the inference.

I have the impression that I now investigate genuine properties of the model
and its failures to describe the data. I have previously seen that the model
has a hard time generating noise in early stages of the decision making
process. So the current model may just not be capable to represent sufficiently
large variability in RTs to describe the data.

In all this I don't like that the inference identifies large variability of
bias as a solution. It doesn't make sense: each posterior sample with large
bias compared to the bound makes an immediate decision with a 50% accuracy.
This should be strongly penalised, but it isn't. Perhaps here the large number
of RT quantiles overwhelms the single accuracy statistics for the choices. The
inference probably trades off two evils against each other at this stage. I
could reduce the number of RT quantiles in the summary statistics, but the
issue appears to simply be that there is no good set of parameter values that
fits to the observed data.