alternative-specific quantiles as summary statistics
I realised that the wide posterior for the bias, i.e., the possibility of
having large biases as solution could be attributed to the summary statistics:
Because I only used quantiles across all choices, the inference didn't have to
care about the choices for getting the quantiles right. To counter this I now
use two sets of quantiles, one per alternative in the summary statistics. This
prevents inference from using large bias variability as solution, because large
biases will always choose one alternative making the RT quantiles of the other
alternative all 0 (in my implementation).

The inference results don't satisfy me, however. Even with fixed intstd=0.4
(which was a reasonable value before), the quantile fit is not so good. This is
because the posterior is bimodal, especially in noisestd and bound. There is
one solution with bound and noisestd very low and one with bound~0.51 and
noisestd~0.28, but even for this more reasonable solution decisions are made
almost immediately with most of RT explained by non-decision time.

name         bound    lapseprob       ndtloc     noisestd
5%        0.500268     0.024942    -1.437067     0.045740
50%       0.507531     0.042021    -1.255228     0.205387
95%       0.523734     0.071619    -1.141129     0.337896

I don't know why the low bound and noisestd solution is found at all, because
this appears to induce very large variability in the prediction of the summary
statistics leading to many outliers in the boxplot. The medians indicated by
the boxplots, however, lie quite close to data RT quantiles suggesting that at
least the one solution is reasonable (apart from the high non-decision time).

In conclusion, the summary statistics remains an issue. I'm thinking now that
I should somehow scale the RT quantiles of the individual alternatives with
their choice frequency instead of using the accuracy as an additional
dimension. This would work more straightforward with the histogram summary
statistics. Perhaps I should also let the DDM approaches inspire me. At least
they were using something like summary statistics for fitting all along. Still,
my preferred approach would be to estimate the original likelihood directly,
not that of some summary statistics!