new quantile-based histogram summary statistics
In search for a sensible summary statistics I checked what the DDM people have
been doing. Two of the approaches they describe (Ratcliff2002,
Vandekerckhove2007) are based on a multinomial and a quantile-based chi-square
statistic. Both of them boild down to estimating frequencies of responses
within RT bins defined by the data RTs. This makes a lot of sense for me, too:
By defining histogram bins based on RT quantiles, you automatically define bins
located around the data. By counting the frequency of a response within RT bins
you get information about the responses into the summary statistics.

I have now reformulated Stats_hist to implement this summary statistics. I ran
two repetitions of inference on the data of subject 19 with intstd fixed to
0.4. Both gave roughly consistent results with one estimating slightly higher
noisestd values:

name         bound    lapseprob       ndtloc     noisestd
5%        0.525611     0.042712    -1.650439     0.560033
50%       0.540962     0.117674    -1.435304     0.788637
95%       0.558400     0.310149    -1.246229     0.970351

name         bound     lapseprob       ndtloc     noisestd
5%        0.531702  4.844519e-02    -1.746384     0.577613
50%       0.547150  1.213162e-01    -1.544369     0.722620
95%       0.560034  2.322256e-01    -1.330462     0.828714

In both results I see some larger posterior correlations, but they all stay
below a magnitude of 0.7 and don't appear to spread the marginal posterior
excessively. The posterior predictive diagnostics also look quite similar
across the two repetitions. Specifically, both suggest a reasonable tail of
decision times reaching beyond 0.5 s. The boxplot of summary statistic values
for the posterior samples indicate a reasonable fit to the ones from the data
with most data values falling within 25% - 75% of the posterior values.
However, the spread of the posterior values can extend considerably beyond the
data value (e.g. data = 0.08, posterior sample = 0.4 which would indicate many
more responses in the corresponding quantile bin).

The posteriors now contain a sizable amount of lapse probability again
(~0.1-0.2). The model needs this to explain some data in the tail of the RT
distribution and is not strongly penalised for it, because the bins all contain
reasonable amounts of data. Perhaps I could remove the lapse probability
explanation by including a bin not containing data, e.g., [0, 0.1].

Full model inference
--------------------
I also tested how inference works when I fit intstd together with the other
parameters. This led to a solution with a greater contribution of non-decision
time to RTs. intstd and bound showed the expected correlation in the posterior
which led to a large spread especially of bound values which covered the whole
range [0.5, 1], but also bias had increased spread compared to the constrained
model.

name         bound       intstd    lapseprob       ndtloc  noisestd
5%        0.524001     0.087468     0.000840    -1.554378  0.295390
50%       0.672991     0.194786     0.023174    -1.304519  0.753866
95%       0.924074     0.417344     0.170323    -1.086193  1.262896

The summary statistic appears to be fitted quite well, although the posterior
values are even more spread out than for the constrained model. The posterior
predictive plot of the RT distributions indicates a reasonably good fit, but it
looks slightly worse than those from the constrained model with specifically
problems with the tails of the response distribution.

Overall, this fit also appears to be good, but I'm inclined to fix the bound
now as the data appears to not constrain it much at all while intstd stays at
relatively low values. Perhaps this is also because I have a uniform prior on
the bound, but the prior of intstd biases it towards 0. On the other hand, the
median of the prior is at 1 which is considerably beyond the estimated
posterior values.