check difficulty differences for first full set of snl inference results
As the most interesting statistic I was first interested in the differences in
accuracy and median RT between the two difficulty levels across subjects. So
this script computes this based on the first full set of inference results
which can be found in folder:

	snl/rotated_directions/201807061824

Specifically, the script re-creates the models used during inference and then
uses the stored posterior samples to compute posterior predictive estimates of
the accuracy and median RT for the two difficulties. This takes a while (~10 to
15 min for the 36 subjects).

Note that reading the HDF-store did not work in my standard python3
environment, but it worked ok in the python 2 based snl environment that
generated the results.

Results
-------
As the results for subject 19 have already hinted at, the model really has a
problem with representing the differences in accuracy across difficulties. In
most subjects the model estimated this difference to be around 0.13 or so
although this value is below 0.1 in the data for most subjects. Also, there
doesn't appear to be a dependence of the posterior predictive accuracy
differences on the corresponding data values.

The median RT differences are reasonably well represented in the posterior
predictive estimates, but these also have a relatively large spread compared to
the values themselves (e.g., median RT difference in data of -0.02 with
posterior predictive values roughly within [-0.12, 0.05] with most values
within [-0.04, 0])

Overall, the model appears to have serious issues with representing the
difficulty differences of these summary measures. I don't think that this is a
problem with the inference, but a property of the model that needs to be
investigated.