alternative analysis of behavioural data with HDDM
When I still had problems with the right summary statistics for snl-inference,
Dragan suggested to check whether we could fit the data with the DDM. I was
also interested to see whether the DDM would be able to represent the accuracy
and median RT differences better than our perceptual model. It was clear from
the start, however, that the DDM has much better chances to do that, because it
would use different parameters for the different difficulties so that it is the
more complex model.

It was easy to implement this inference with the HDDM (version 0.6.0 in Python
3). I decided to use a DDM model very close to my perceptual model, i.e., I
included a lapse probability (p_outlier) although I fixed this to 0.05 during
inference. I also used stimulus encoding which allowed me to include a bias
parameter. I did not use HDDM's prime feature of doing group inference, but
fitted individual subjects independently. Given the large number of trials this
should be fine. The biggest difference to the perceptual model is that I
allowed drift to vary with difficulty so that the model could pick up
differences across difficulty. Further, I used a constant non-decision time
instead of a uniformly distributed ndt as in the perceptual model. I have not
tried alternative parameterisations so far.

Other than for snl, I directly generate and store posterior predictive samples
during inference so that I don't have to re-create the fitted HDDM model for
analysis.

Results
-------
are in
	hddm_result_201807101804.h5

As expected, the HDDM represents the difficulty differences in accuracy and
median RT better than the perceptual model. Especially the accuracy values are
fitted much better and show a clear relation to the data values although the
HDDM also tends to slightly overestimate the accuracy differences, when they
are small.

I also checked correlations between the estimated drift rate difficulty
differences and the accuracy and median RT difficulty differences. These
indicate the expected relationship of higher drift differences for higher
accuracy and lower median RT differences, but the relation is noisier than I
had expected.

My conclusion from this comparative analysis is that the perceptual model has
some principled issues with representing the data. In theory it could still be
an issue with the inference. I could exclude this possibility by implementing a
DDM simulator and let snl do the inference for the DDM, too, but my intuition
is that the issue is with the model, not the inference. Otherwise I don't
currently have any meaningful insights from the HDDM analysis.