testing inference with model time resolution of 1 ms instead of 5 ms
I tested this only for subject 19. Results are in:

	snl/rotated_directions/201808081908

As expected the computation time increased; from 35 min to 52 min. The
inference results are:

name        intstd     lapseprob       ndtloc     noisestd
5%        0.124676  5.236301e-04    -1.530506     0.728582
50%       0.156965  5.986315e-02    -1.309160     1.192230
95%       0.187485  2.280600e-01    -1.072615     1.855876

compared to the results of commit 70fbe684 intstd, lapseprob and ndtloc have
very similar posterior distributions, just that they are a bit wider. noisestd,
however, is considerably larger now (median moved from 0.77 to 1.19). However,
I also just realised that I here used bound=0.7 while I previously used 0.8. So
it is unclear how much the increase in noisestd was driven by the lower dt or
by the lower bound.

I do not have quantitative comparisons available, but the posterior predictive
checks appear to give very similar results to before. Specifically, the
difficulty differences in accuracy and median RT exhibit the previously
observed deficiancies: The posterior predictive estimates of the accuracy
difference were mostly within [0.08, 0.13] (edges of box) which is similar to
before (with a data value of 0.03). Posterior predictive median RT differences
were mostly in [-0.03, 0] which is also similar to before.

Although the increase in inferred noisestd values is a bit curious, overall the
inference results, especially the posterior predictive checks, were very
similar to before. I conclude from this that the time resolution of the model
is not the crucial factor preventing the model from better explaining the
accuracy and median RT differences in the data.

Warning
-------
Should I ever want to manipulate the time resolution dt more regularly I need
to save it together with the results. So far it's hard-coded in helpers.py.