realising that criteria+difference combinations allow 2 directions
During implementation of the model I realised that any combination of criterion
and difference does not uniquely determine the underlying motion direction.
Instead, each combination defines two possible motion directions corresponding
to the criterion as either a down or up direction (or left and right direction,
respectively). I now account for that fact in the model definition.