first working implementation of the difference model
Following the model definition and using the existing rotated directions code
as template, I created an implementation of the difference model. This
implementation now encodes angles in integer degrees and only uses conversion
to rad when interacting with continuous density functions. I did this to
simplify the interface and make it more human-interpretable. Further cleanup of
the interface included putting all trial-dependent data into attribute
'Trials' (meaning that I moved data that was previously stored in criteria into
Trials and use criteria now to hold the encoded set of criterion values).

It turned out that the correct implementation of the interactions between
decision-defining differences, criterion and motion direction values was not
trivial, mainly because the criterion always has two direction interpretations
so that computations with the criterion always involve two angles. I believe I
have solved this properly now.

For the implementation in numba I had to adapt a few things from my original
Python implementation of the evidence accumulation process:
- numba doesn't support array expansion functions such as 'tile' so I had to
  implement this with for-loops (which is not bad for a compiled function)
- numba doesn't support np.r_ or np.c_ - I replaced occurrences with np.stack
- the biggest headache was that the parallel for-loop did not work, I cannot
  tell what eventually made it work, but it could have been something related
  to using -= or += operators for the wrong variable
With the parallel for-loop working, simulations from the model are pretty fast.
I previously also used target='parallel' for compute_minimal_differences, but
this appeared to produce more overhead than time savings.

Initial tests
-------------
Produced promising results. RT distributions obviously strongly depend on the
chosen parameters. So far I have verified that
- without noise the model always makes the correct choice
- usually easy decisions are faster than hard decisions

But there are also some weird effects. The weirdest is that for large
critstd=0.45, i.e., prior and therefore also posterior uncertainty over the
criterion, the RT distributions in easy and hard trials become bimodal and a
sizable fraction of easy trials actually are timed out while this doesn't
happen in hard trials. The issue here appears to be that information coming
from the estimate of the criterion is successively overwritten by increasingly
certain information about the motion direction. I investigated this by plotting
lE_DR over time points within noise-free trials. This stores the log-evidence
for difference-criterion combinations which is then used to infer the most
likely difference value from which the decision is determined. Initially the
values in lE_DR differ only due to the estimated criterion values which favour
the true criterion. In each time step we then add evidence for
difference-criterion combinations coming from the observed motion direction.
The point is that initially the evidence for differences may favour a clockwise
choice, but that the increasing influence of the motion directions may later
induce evidences that slightly favour an anti-clockwise choice, when averaged
across possible criterion values.

Here's an example: true criterion=45, true direction=15 -> true difference=30
So the difference-criterion combination 30x45 should have the highest evidence
throughout the decision process and this is the case. More interesting is the
combination -30x45 which represents the directions 75 and 255. This combination
has initially one of the largest evidence, because it uses the correct
criterion. But because the directions it represents are 60 and 120 degrees away
from the true direction, it accumulates direction evidence slowly. The opposite
is true for the combination -30x135 which represents directions 165 and 345.
Because this combination represents a criterion value very different from the
true one, it initially has one of the lowest evidences, but because the motion
directions it represents are only 30 and 180 degrees away, it accumulates
evidence faster through time (apparently the directions that are over 100
degrees away from the true one don't contribute anything to the evidences in
this example). Consequently, the evidence for -30x135 overtakes that for -30x45
after about 30 time steps. Because both of them represent anti-clockwise
rotations, this does not affect the decision, but across all combinations such
effects apparently add up and then produce the strange behaviour that evidence
for clockwise decreases again after some time. Note that lE_DR needs to be
logsumexp-ed over the criterion dimension to get evidence for differences.
Summing directly actually leads to equal evidence for positive and negative
differences. Parameters used in this example:

model.diffstd = 1
model.critstd = 0.45
model.dirstd = 0.15
model.cnoisestd = 1e-12
model.dnoisestd = 1e-12

So this basically explains how it can happen that evidence for one alternative
stagnates after some time. I still find it odd, however, that this only occurs
for easy decisions with differences of +/-30 degrees and not for hard decisions
with differences of +/-15 degrees. I tried to check this. For the example above
I looked for trials with the same motion direction, but as a hard decision.
There were only hard trials with true direction=15 and true criterion=0 in the
experiment. In these trials the effect cannot occur, because from the beginning
of the trial evidence already points to anti-clockwise and the effect of the
motion direction evidence was that it increasingly pointed towards
anti-clockwise meaning that it cannot lead to a stagnation, or reversal of the
evidence accumulation process. So I instead used true criterion=30 which did
not occur in the experiment in simulation. Here, I saw the effect also for the
hard decision: evidence for clockwise rotation increased initially, but then
reduced again after 20 time steps and consequently never reached the decision
bound. I, therefore, believe that it could be a coincidence resulting from the
particular experimental design that the effect only occurs in easy decisions.